api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
tortoise-orm==0.23.0
asyncpg==0.30.0
aerich==0.8.1
tomlkit==0.13.2
redis==5.2.1 
celery==5.4.0
pydantic==2.10.5
bcrypt==4.0.1
passlib==1.7.4 
python-jose==3.3.0
python-multipart==0.0.20